refactor(interaction): extract the scroll command runtime out of gestures.ts
